# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "High Street Gym ",
    "description": "JSON REST API for Dynamic Gym website "
  },
  "servers": [
    {
      "url": "http://localhost:8090"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/blog_post": {
      "post": {
        "summary": "Create a blog_post",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new blog_post",
          "content": {
            "application/json": {
              "schema": {
                "blog_post_user_id": "number",
                "blog_post_class_id": "number",
                "blog_post_created_date_time": "string"
              },
              "example": {
                "blog_post_user_id": 1,
                "blog_post_class_id": 1,
                "blog_post_created_date_time": "2023/10/20"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all blog_post",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Update an existing blog_post",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update an existing blog_post",
          "content": {
            "application/json": {
              "schema": {
                "blog_post_user_id": "number",
                "blog_post_class_id": "number",
                "blog_post_created_date_time": "string",
                "blog_post_id": "number"
              },
              "example": {
                "blog_post_user_id": 110,
                "blog_post_class_id": 110,
                "blog_post_created_date_time": "2023/04/10",
                "blog_post_id": 1
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific blog_post by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deleting a blog_post",
          "content": {
            "application/json": {
              "schema": {
                "blog_post_id": "number"
              },
              "example": {
                "blog_post_id": "1"
              }
            }
          }
        }
      }
    },
    "/blog_post/{id}": {
      "get": {
        "summary": "Get a specific blog_post by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}